CLN: create image_tab module
